Issue #27
Bugfix: Integer editor allows non numeric characters.

Solution: It was actually an internal NumberFormatException thrown by
the JavaFX Spinner class. It uses by default the IntegerStringConverter.
However that converter will throw a NumberFormatException when the
string can not be parsed to an int. Therefore added a new class
"NoExceptionStringConverter" which will return a default value when the
string can not be parsed instead of throwing an exception and leaving
invalid text inside the editor. For now it returns 0.

Also added "KeyListener" which will also set a default value "0" when
the enter key is pressed and the string can not be parsed.