convert spinner value when defocusing or pressing the tab key
When only converting the spinner value to remove non-integer characters upon pressing enter, clicking outside of the spinner or pressing tab will not result in the value being converted. Added another conversion when the spinner is being defocused, which is the case when tab is pressed or the mouse is being clicked to something different.